requirements.txt
pandas==1.1.3
gunicorn==20.0.4
flask-api==2.0
Flask==2.0.3
Jinja2==3.0.3
google-cloud-bigquery==2.20.0
numpy==1.21.2
scipy==1.7.1
pyarrow==5.0.0 